Made the app even MORE modular:
* Moved helper functions
* Added some more error checking
